feat: Postman에 맞는 api주소 넣기, Type 정리
