+++ src/apis/community.ts
@@ -1,7 +1,7 @@
-import { string } from 'prop-types';
 import axios from 'axios';
 import axiosInstance from '@apis/axiosInstance';
-const END_POINT = '/';
+const END_POINT =
+  'https://c69dc827-4460-41db-8879-0e6d753aff12.mock.pstmn.io/api/v1/';
 
 export const getCommunityDataApi = async () => {
   try {
@@ -17,10 +17,11 @@ export const getCommunityDataApi = async () => {
 
 export const deleteCommunityPostApi = async (url: string) => {
   try {
-    await axiosInstance({
+    const res = await axios({
       url: `${END_POINT}${url}`,
       method: 'DELETE',
     });
+    console.log(res);
   } catch (error) {
     new Error('Community Post를 삭제하는데 오류가 발생하였습니다.');
   }
@@ -34,11 +35,12 @@ export const postCommunityPostApi = async ({
   postData: FormData;
 }) => {
   try {
-    await axiosInstance({
+    const res = await axios({
       url: `${END_POINT}${url}`,
       method: 'POST',
       data: postData,
     });
+    console.log(res);
   } catch (error) {
     new Error('Community Post 생성하는데 오류가 발생하였습니다.');
   }
@@ -51,11 +53,12 @@ export const editCommunityPostApi = async ({
   postData: FormData;
 }) => {
   try {
-    await axiosInstance({
+    const res = await axios({
       url: `${END_POINT}${url}`,
       method: 'PATCH',
       data: postData,
     });
+    console.log(res);
   } catch (error) {
     new Error('Community Post 수정하는데 오류가 발생하였습니다.');
   }
@@ -63,10 +66,11 @@ export const editCommunityPostApi = async ({
 
 export const postCommunityPostLikeApi = async (url: string) => {
   try {
-    await axiosInstance({
+    const res = await axios({
       url: `${END_POINT}${url}`,
       method: 'POST',
     });
+    console.log(res);
   } catch (error) {
     new Error('Community Post Like를 등록하는데 오류가 발생하였습니다.');
   }
@@ -74,10 +78,11 @@ export const postCommunityPostLikeApi = async (url: string) => {
 
 export const deleteCommunityPostLikeApi = async (url: string) => {
   try {
-    await axiosInstance({
+    const res = await axios({
       url: `${END_POINT}${url}`,
       method: 'DELETE',
     });
+    console.log(res);
   } catch (error) {
     new Error('Community Post Like를 삭제하는데 오류가 발생하였습니다.');
   }
