+++ src/interfaces/community.d.ts
@@ -17,3 +17,25 @@ export interface CommunityType {
   posts?: CommunityPostType[];
   hasNext?: boolean;
 }
+
+export interface CommunityModalType {
+  postId?: string;
+  nickname?: string;
+  profileImageUrl?: string;
+  modalType: string;
+  isOpen?: boolean;
+  onClose?: (e: React.MouseEvent<HTMLElement>) => void;
+}
+
+interface CommunityPostMenuIconButtonType {
+  postId: string;
+  nickname: string;
+  profileImageUrl?: string;
+}
+
+export interface CommunityPostTypographyButtonType {
+  name?: string | undefined;
+  margin?: string;
+  children?: string | number;
+  onClick?: (e: React.MouseEvent<HTMLElement>) => void;
+}
